Base Component of the program
This is the finalised version of the program with all of the components integrated to create a final product. It asks the user for a recipe name, then allows the user to add ingredients to their recipe whilst the program constantly updates the user on the price and items included. Once the user has finished inputting ingredients they can input 'xxx' to stop adding ingredients and view the final cost and length of their recipe